Adding sender and receiver to posts and full post view for posts. Added sender and receiver names to Dashboard posts and also on click of bigger size image post showing a full post view for respected post.
